The WOW pledge placard moves to its own file, unchanged (#2328)
Pure extraction, no repaint. `WOWSelectCard` leaves
`selectCard/FactionSelectCard.tsx` for `selectCard/WowSelectCard.tsx` as a
default export, matching how `components/taskCard/` is laid out and the pattern
`pickVariant` call, and `factions/wow.ts` registers the new file as a plain
`lazyArchetype` instead of reaching for a named export.

Proven byte-identical: `renderToStaticMarkup` over all six
`state` x `members` combinations, before and after, diffed — 23,287 bytes of
markup, zero difference.

The dispatcher's "the tiles are dark" docstring is NOT carried into the new
file (#2321): the premise is false and it is what let dark-only registers look
correct on a surface that flips.
